Jd/simulation sequence fix

+++ src/devices_models/devices/common/constraints_structs.jl
@@ -12,10 +12,10 @@ end
 
 struct DeviceTimeSeries
     name::String
-    bus_number::Int64
+    bus_number::Int
     multiplier::Float64
     timeseries::Vector{Float64}
-    range::Union{Nothing,DeviceRange}
+    range::Union{Nothing, DeviceRange}
     function DeviceTimeSeries(name, bus_number, multiplier, timeseries, range)
         @assert isnothing(range) || name == range.name
         return new(name, bus_number, multiplier, timeseries, range)
