+++ src/devices_models/devices/common/expressionarray_algebra.jl
@@ -1,6 +1,6 @@
 ###### Operations for JuMPExpressionMatrix ######
 
-function _remove_undef!(expression_array::T) where {T<:JuMPExpressionMatrix}
+function _remove_undef!(expression_array::T) where {T <: JuMPExpressionMatrix}
     for j in 1:size(expression_array)[2]
         for i in 1:size(expression_array)[1]
             if !isassigned(expression_array, i, j)
@@ -23,11 +23,11 @@ end
 
 function _add_to_expression!(
     expression_array::T,
-    ix::Int64,
-    jx::Int64,
+    ix::Int,
+    jx::Int,
     var::JV,
     multiplier::Float64,
-) where {T,JV<:JuMP.AbstractVariableRef}
+) where {T, JV <: JuMP.AbstractVariableRef}
     if isassigned(expression_array, ix, jx)
         JuMP.add_to_expression!(expression_array[ix, jx], multiplier, var)
     else
@@ -39,12 +39,12 @@ end
 
 function _add_to_expression!(
     expression_array::T,
-    ix::Int64,
-    jx::Int64,
+    ix::Int,
+    jx::Int,
     var::JV,
     multiplier::Float64,
     constant::Float64,
-) where {T,JV<:JuMP.AbstractVariableRef}
+) where {T, JV <: JuMP.AbstractVariableRef}
     if isassigned(expression_array, ix, jx)
         JuMP.add_to_expression!(expression_array[ix, jx], multiplier, var)
         JuMP.add_to_expression!(expression_array[ix, jx], constant)
@@ -57,8 +57,8 @@ end
 
 function _add_to_expression!(
     expression_array::T,
-    ix::Int64,
-    jx::Int64,
+    ix::Int,
+    jx::Int,
     value::Float64,
 ) where {T}
     if isassigned(expression_array, ix, jx)
@@ -72,8 +72,8 @@ end
 
 function _add_to_expression!(
     expression_array::T,
-    ix::Int64,
-    jx::Int64,
+    ix::Int,
+    jx::Int,
     parameter::PJ.ParameterRef,
 ) where {T}
     if isassigned(expression_array, ix, jx)
