+++ src/devices_models/devices/common/pm_translator.jl
@@ -1,24 +1,18 @@
 
 struct PMmap
-    bus::Dict{Int64,PSY.Bus}
+    bus::Dict{Int, PSY.Bus}
     arcs::Dict{
-        NamedTuple{
-            (:from_to, :to_from),
-            Tuple{Tuple{Int64,Int64,Int64},Tuple{Int64,Int64,Int64}},
-        },
-        t where t<:PSY.ACBranch,
+        NamedTuple{(:from_to, :to_from), Tuple{Tuple{Int, Int, Int}, Tuple{Int, Int, Int}}},
+        t where t <: PSY.ACBranch,
     }
     arcs_dc::Dict{
-        NamedTuple{
-            (:from_to, :to_from),
-            Tuple{Tuple{Int64,Int64,Int64},Tuple{Int64,Int64,Int64}},
-        },
-        t where t<:PSY.DCBranch,
+        NamedTuple{(:from_to, :to_from), Tuple{Tuple{Int, Int, Int}, Tuple{Int, Int, Int}}},
+        t where t <: PSY.DCBranch,
     }
 end
 
-function get_branch_to_pm(ix::Int64, branch::PSY.PhaseShiftingTransformer)
-    PM_branch = Dict{String,Any}(
+function get_branch_to_pm(ix::Int, branch::PSY.PhaseShiftingTransformer)
+    PM_branch = Dict{String, Any}(
         "br_r" => PSY.get_r(branch),
         "rate_a" => PSY.get_rate(branch),
         "shift" => PSY.get_α(branch),
@@ -41,8 +35,8 @@ function get_branch_to_pm(ix::Int64, branch::PSY.PhaseShiftingTransformer)
     return PM_branch
 end
 
-function get_branch_to_pm(ix::Int64, branch::PSY.Transformer2W)
-    PM_branch = Dict{String,Any}(
+function get_branch_to_pm(ix::Int, branch::PSY.Transformer2W)
+    PM_branch = Dict{String, Any}(
         "br_r" => PSY.get_r(branch),
         "rate_a" => PSY.get_rate(branch),
         "shift" => 0.0,
@@ -65,8 +59,8 @@ function get_branch_to_pm(ix::Int64, branch::PSY.Transformer2W)
     return PM_branch
 end
 
-function get_branch_to_pm(ix::Int64, branch::PSY.TapTransformer)
-    PM_branch = Dict{String,Any}(
+function get_branch_to_pm(ix::Int, branch::PSY.TapTransformer)
+    PM_branch = Dict{String, Any}(
         "br_r" => PSY.get_r(branch),
         "rate_a" => PSY.get_rate(branch),
         "shift" => 0.0,
@@ -89,8 +83,8 @@ function get_branch_to_pm(ix::Int64, branch::PSY.TapTransformer)
     return PM_branch
 end
 
-function get_branch_to_pm(ix::Int64, branch::PSY.Line)
-    PM_branch = Dict{String,Any}(
+function get_branch_to_pm(ix::Int, branch::PSY.Line)
+    PM_branch = Dict{String, Any}(
         "br_r" => PSY.get_r(branch),
         "rate_a" => PSY.get_rate(branch),
         "shift" => 0.0,
@@ -113,8 +107,8 @@ function get_branch_to_pm(ix::Int64, branch::PSY.Line)
     return PM_branch
 end
 
-function get_branch_to_pm(ix::Int64, branch::PSY.HVDCLine)
-    PM_branch = Dict{String,Any}(
+function get_branch_to_pm(ix::Int, branch::PSY.HVDCLine)
+    PM_branch = Dict{String, Any}(
         "loss1" => PSY.get_loss(branch).l1,
         "mp_pmax" => PSY.get_reactivepowerlimits_from(branch).max,
         "model" => 2,
@@ -147,21 +141,15 @@ function get_branch_to_pm(ix::Int64, branch::PSY.HVDCLine)
 end
 
 function get_branches_to_pm(sys::PSY.System)
-    PM_ac_branches = Dict{String,Any}()
-    PM_dc_branches = Dict{String,Any}()
+    PM_ac_branches = Dict{String, Any}()
+    PM_dc_branches = Dict{String, Any}()
     PMmap_ac = Dict{
-        NamedTuple{
-            (:from_to, :to_from),
-            Tuple{Tuple{Int64,Int64,Int64},Tuple{Int64,Int64,Int64}},
-        },
-        t where t<:PSY.ACBranch,
+        NamedTuple{(:from_to, :to_from), Tuple{Tuple{Int, Int, Int}, Tuple{Int, Int, Int}}},
+        t where t <: PSY.ACBranch,
     }()
     PMmap_dc = Dict{
-        NamedTuple{
-            (:from_to, :to_from),
-            Tuple{Tuple{Int64,Int64,Int64},Tuple{Int64,Int64,Int64}},
-        },
-        t where t<:PSY.DCBranch,
+        NamedTuple{(:from_to, :to_from), Tuple{Tuple{Int, Int, Int}, Tuple{Int, Int, Int}}},
+        t where t <: PSY.DCBranch,
     }()
 
     for (ix, branch) in enumerate(PSY.get_components(PSY.Branch, sys))
@@ -186,10 +174,10 @@ function get_branches_to_pm(sys::PSY.System)
 end
 
 function get_buses_to_pm(buses::IS.FlattenIteratorWrapper{PSY.Bus})
-    PM_buses = Dict{String,Any}()
-    PMmap_buses = Dict{Int64,PSY.Bus}()
+    PM_buses = Dict{String, Any}()
+    PMmap_buses = Dict{Int, PSY.Bus}()
 
-    pm_bustypes = Dict{PSY.BusType,Int64}(
+    pm_bustypes = Dict{PSY.BusType, Int}(
         PSY.ISOLATED => 4,
         PSY.PQ => 1,
         PSY.PV => 2,
@@ -199,7 +187,7 @@ function get_buses_to_pm(buses::IS.FlattenIteratorWrapper{PSY.Bus})
 
     for bus in buses
         number = PSY.get_number(bus)
-        PM_bus = Dict{String,Any}(
+        PM_bus = Dict{String, Any}(
             "zone" => 1,
             "bus_i" => number,
             "bus_type" => pm_bustypes[PSY.get_bustype(bus)],
@@ -222,22 +210,22 @@ function get_buses_to_pm(buses::IS.FlattenIteratorWrapper{PSY.Bus})
     return PM_buses, PMmap_buses
 end
 
-function pass_to_pm(sys::PSY.System, time_periods::Int64)
+function pass_to_pm(sys::PSY.System, time_periods::Int)
 
     ac_lines, dc_lines, PMmap_ac, PMmap_dc = get_branches_to_pm(sys)
     buses = PSY.get_components(PSY.Bus, sys)
     pm_buses, PMmap_buses = get_buses_to_pm(buses)
-    PM_translation = Dict{String,Any}(
+    PM_translation = Dict{String, Any}(
         "bus" => pm_buses,
         "branch" => ac_lines,
         "baseMVA" => sys.basepower,
         "per_unit" => true,
-        "storage" => Dict{String,Any}(),
+        "storage" => Dict{String, Any}(),
         "dcline" => dc_lines,
-        "gen" => Dict{String,Any}(),
-        "switch" => Dict{String,Any}(),
-        "shunt" => Dict{String,Any}(),
-        "load" => Dict{String,Any}(),
+        "gen" => Dict{String, Any}(),
+        "switch" => Dict{String, Any}(),
+        "shunt" => Dict{String, Any}(),
+        "load" => Dict{String, Any}(),
     )
 
     # TODO: this function adds overhead in large number of time_steps
