+++ test/test_utils/model_checks.jl
@@ -1,14 +1,14 @@
-const GAEVF = JuMP.GenericAffExpr{Float64,VariableRef}
-const GQEVF = JuMP.GenericQuadExpr{Float64,VariableRef}
+const GAEVF = JuMP.GenericAffExpr{Float64, VariableRef}
+const GQEVF = JuMP.GenericQuadExpr{Float64, VariableRef}
 
 function moi_tests(
     op_problem::OperationsProblem,
     params::Bool,
-    vars::Int64,
-    interval::Int64,
-    lessthan::Int64,
-    greaterthan::Int64,
-    equalto::Int64,
+    vars::Int,
+    interval::Int,
+    lessthan::Int,
+    greaterthan::Int,
+    equalto::Int,
     binary::Bool,
 )
